edge/speech_recognition/deepspeech2_pytorch/requirements.txt
jupyterlab
librosa
numpy
psutil
pysoundfile
pytest
python_speech_features
requests
tensorboardX
tensorflow

https://download.pytorch.org/whl/cu91/torch-0.4.0-cp35-cp35m-linux_x86_64.whl
torchvision
